requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-text-splitters==0.0.1
langchain-openai==0.0.2
chromadb==0.4.22
python-dotenv==1.0.0
flask==3.0.0
PyPDF2==3.0.1
openai==1.6.1